requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
async-generator==1.10
attrs==21.4.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
cryptography==36.0.1
cssselect2==0.6.0
decorator==5.1.0
distlib==0.3.3
Django==4.0
django-cors-headers==3.10.1
django-crontab==0.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.3.0
future==0.18.2
h11==0.13.0
html5lib==1.1
idna==3.3
lxml==4.9.0
numpy==1.22.2
oscrypto==1.3.0
outcome==1.1.0
pandas==1.4.1
Pillow==9.1.1
platformdirs==2.4.0
psutil==5.8.0
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pyOpenSSL==21.0.0
PyPDF3==1.0.6
pyte==0.8.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
win-unicode-console==0.5
wsproto==1.0.0
xhtml2pdf==0.2.7
